+++ app/src/main/java/com/michaeltroger/gruenerpass/pdf/QrPagerFragment.kt
@@ -1,0 +1,28 @@
+package com.michaeltroger.gruenerpass.pdf
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import androidx.fragment.app.Fragment
+import com.michaeltroger.gruenerpass.R
+
+class QrPagerFragment(private val pdfHandler: PdfHandler) : Fragment() {
+
+    private var qrCode: ImageView? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View = inflater.inflate(R.layout.layout_item_qr, container, false)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        qrCode = view.findViewById(R.id.qrcode)
+        qrCode?.setImageBitmap(pdfHandler.getQrBitmap())
+    }
+
+}
